Refactor and improve code for emissions treemap
This commit reuses existing code for geo,  emissions and population
(after a minor refactoring). It also simplifies code for merging the
datasets and for plotting the graphs.

The treemap graph code gets overhauled (basically by copying code from squarify.plot and customizing it for the needs of this plot). It now:
- supports flipping the y-axis,
- adds minor padding,
- resizes the plot so that most labels fit in,
- adds emissions_summary on another line.